Due to the fact that a new method now existed within the EMF class, the endpoints (Account,Cycle,CycleUse,ResourcePurchase and Transaction Log) had to be refactored to use this new method which was getManagerInstance().
It was noted that the closing of the entity manager a the end of methods in the account endpoint produced an error so it was removed throughout.

In terms of updating an account, changes should only be made to the country, address and county. In the case of change in name etc , the account would need to be deleted.